+++ lib/harmoniser/launcher/bounded.rb
@@ -1,11 +1,8 @@
-require "harmoniser/channelable"
 require_relative "base"
 
 module Harmoniser
   module Launcher
     class Bounded < Base
-      include Channelable
-
       private
 
       def start_subscribers
@@ -15,7 +12,7 @@ module Harmoniser
       end
 
       def channel
-        @channel ||= self.class.create_channel(consumer_pool_size: @configuration.concurrency, consumer_pool_shutdown_timeout: @configuration.timeout)
+        @channel ||= Subscriber.create_channel(consumer_pool_size: @configuration.concurrency, consumer_pool_shutdown_timeout: @configuration.timeout)
       end
     end
   end